relocate single proplist dialogs
