+++ MFCMapi.vcxproj
@@ -631,7 +631,7 @@ copy "$(TargetDir)$(TargetName).pdb" "$(SolutionDir)"
     <ClInclude Include="UI\Dialogs\MFCUtilityFunctions.h" />
     <ClInclude Include="UI\Dialogs\Editors\QSSpecialFolders.h" />
     <ClInclude Include="UI\RichEditOleCallback.h" />
-    <ClInclude Include="UI\Dialogs\SingleRecipientDialog.h" />
+    <ClInclude Include="UI\Dialogs\propList\SingleRecipientDialog.h" />
     <ClInclude Include="UI\ViewPane\CheckPane.h" />
     <ClInclude Include="UI\ViewPane\CountedTextPane.h" />
     <ClInclude Include="UI\ViewPane\DropDownPane.h" />
@@ -656,7 +656,7 @@ copy "$(TargetDir)$(TargetName).pdb" "$(SolutionDir)"
     <ClInclude Include="UI\Dialogs\ContentsTable\RecipientsDlg.h" />
     <ClInclude Include="UI\Dialogs\Editors\RestrictEditor.h" />
     <ClInclude Include="UI\Dialogs\ContentsTable\RulesDlg.h" />
-    <ClInclude Include="UI\Dialogs\SingleMessageDialog.h" />
+    <ClInclude Include="UI\Dialogs\propList\SingleMessageDialog.h" />
     <ClInclude Include="UI\Controls\SortList\SingleMAPIPropListCtrl.h" />
     <ClInclude Include="UI\Dialogs\Editors\StreamEditor.h" />
     <ClInclude Include="UI\Dialogs\Editors\TagArrayEditor.h" />
@@ -724,8 +724,8 @@ copy "$(TargetDir)$(TargetName).pdb" "$(SolutionDir)"
     <ClCompile Include="UI\RichEditOleCallback.cpp" />
     <ClCompile Include="UI\Dialogs\ContentsTable\RulesDlg.cpp" />
     <ClCompile Include="UI\Controls\SortList\SingleMAPIPropListCtrl.cpp" />
-    <ClCompile Include="UI\Dialogs\SingleMessageDialog.cpp" />
-    <ClCompile Include="UI\Dialogs\SingleRecipientDialog.cpp" />
+    <ClCompile Include="UI\Dialogs\propList\SingleMessageDialog.cpp" />
+    <ClCompile Include="UI\Dialogs\propList\SingleRecipientDialog.cpp" />
     <ClCompile Include="UI\Dialogs\Editors\StreamEditor.cpp" />
     <ClCompile Include="UI\Dialogs\Editors\TagArrayEditor.cpp" />
     <ClCompile Include="UI\UIFunctions.cpp" />
