+++ UI/Dialogs/propList/SingleMessageDialog.h
